General information about starting a group at DIF
These steps provide general context related to #35, describing the overall process for proposing and starting a group. This will need to be revised over time to reference additional details. as we work through our documentation backlog.

As such, this is not perfect at the moment; it's a starting point baseline.